@@ -1,3 +1,4 @@
 export 'package:maxi_library/src/error_handling/error_codes.dart';
 export 'package:maxi_library/src/error_handling/negative_result.dart';
-export 'package:maxi_library/src/error_handling/negative_result_property.dart';
+export 'package:maxi_library/src/error_handling/negative_result_value.dart';
+export 'package:maxi_library/src/error_handling/negative_result_entity.dart';
